fix: check if rollback callback is defined
Don't raise exceptions where a simple print() will suffice
Print exception before prompting for rollback

(cherry picked from commit 4b2fde1d8f3b44a4ee01848549c94f7ce4f224f2)

+++ frappe/installer.py
@@ -91,7 +91,8 @@ def _new_site(
 		enable_scheduler = False
 
 	make_site_dirs()
-	rollback_callback.add(lambda: shutil.rmtree(frappe.get_site_path()))
+	if rollback_callback:
+		rollback_callback.add(lambda: shutil.rmtree(frappe.get_site_path()))
 
 	with filelock("bench_new_site", timeout=1):
 		install_db(
@@ -184,7 +185,8 @@ def install_db(
 
 	if setup:
 		setup_database(force, verbose, no_mariadb_socket)
-		rollback_callback.add(drop_user_and_database(db_name, db_user or db_name))
+		if rollback_callback:
+			rollback_callback.add(drop_user_and_database(db_name, db_user or db_name))
 
 	bootstrap_database(
 		verbose=verbose,
